backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
numpy>=1.23
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
python-dotenv>=1.0
openai>=1.0
langchain>=0.1
langchain-google-genai>=0.1
